requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==4.2.10
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.27.1
importlib-metadata==7.0.1
importlib-resources==6.1.2
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
pkgutil-resolve-name==1.3.10
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
typing-extensions==4.9.0
uritemplate==4.1.1
zipp==3.17.0
